Refactor Battery API code
The JS code written to demo the battery API was written 4 years ago and a lot of things have changed.
The battery API makes use of promises and I changed the old code to reflect the new changes. ES6 arrow
function feature is also included. Also the battery API now works in Chrome not only Firefox